+++ packages/article-byline/article-byline-proptypes.js
@@ -1,12 +1,10 @@
 import PropTypes from "prop-types";
-import { Text, ViewPropTypes } from "react-native";
+import { Text } from "react-native";
 import { treePropType } from "@times-components/markup";
 
 export const articleBylinePropTypes = {
   ast: PropTypes.arrayOf(treePropType).isRequired,
   style: PropTypes.shape({
-    container: ViewPropTypes.style,
-    byline: Text.propTypes.style,
     link: Text.propTypes.style
   })
 };