+++ packages/article-image/article-image.js
@@ -1,14 +1,12 @@
 import React from "react";
 import { View, StyleSheet } from "react-native";
-import Image from "@times-components/image";
-import Caption from "@times-components/caption";
+import ArticleImage from "./component";
 
 import {
   articleImagePropTypes,
   articleImageDefaultPropTypes
 } from "./article-image-proptypes";
 
-// NOTE: We support only Mobile web, ios and android. Desktop to follow after responsive design
 const styles = StyleSheet.create({
   primaryContainer: {
     width: "100%",
@@ -23,14 +21,6 @@ const styles = StyleSheet.create({
     paddingLeft: 10,
     paddingRight: 10
   },
-  secondaryImage: {
-    width: "50%"
-  },
-  secondaryCaption: {
-    paddingLeft: 10,
-    paddingTop: 0,
-    width: "50%"
-  },
   inlineContainer: {
     width: "100%",
     flexDirection: "row",
@@ -38,59 +28,20 @@ const styles = StyleSheet.create({
     paddingBottom: 25,
     paddingLeft: 10,
     paddingRight: 10
-  },
-  inlineImage: {
-    width: "50%"
-  },
-  inlineCaption: {
-    width: "50%",
-    paddingLeft: 10,
-    paddingTop: 0
   }
 });
 
-const captionStyle = {
-  secondary: {
-    container: {
-      paddingTop: 0
-    }
-  },
-  inline: {
-    container: {
-      paddingTop: 0
-    }
-  }
-};
-
-const renderCaption = (display, caption, credits) => {
-  if (!caption && !credits) {
-    return null;
-  }
-  return (
-    <View style={styles[`${display}Caption`]}>
-      <Caption text={caption} credits={credits} style={captionStyle[display]} />
-    </View>
-  );
-};
-
-const ArticleImage = ({ imageOptions, captionOptions }) => {
-  const { display, ratio, url } = imageOptions;
-  const { caption, credits } = captionOptions;
-
-  const [ratioWidth, ratioHeight] = ratio.split(":");
-  const aspectRatio = ratioWidth / ratioHeight;
+const ArticleImageNative = props => {
+  const { display, url } = props.imageOptions;
 
   return (
     <View key={url} style={styles[`${display}Container`]}>
-      <View style={styles[`${display}Image`]}>
-        <Image uri={url} aspectRatio={aspectRatio} />
-      </View>
-      {renderCaption(display, caption, credits)}
+      <ArticleImage {...props} />
     </View>
   );
 };
 
-ArticleImage.propTypes = articleImagePropTypes;
-ArticleImage.defaultProps = articleImageDefaultPropTypes;
+ArticleImageNative.propTypes = articleImagePropTypes;
+ArticleImageNative.defaultProps = articleImageDefaultPropTypes;
 
-export default ArticleImage;
+export default ArticleImageNative;