+++ packages/article/__tests__/web/article.web.test.js
@@ -1,5 +1,4 @@
 /* eslint-env jest */
-
 import shared from "../shared";
 
 describe("Article test on web", () => {