+++ packages/date-publication/date-publication.stories.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { Text, StyleSheet } from "react-native";
 import { storiesOf } from "dextrose/storiesOfOverloader";
 import DatePublication from "./date-publication";
 
@@ -7,6 +8,17 @@ const props = {
   publication: "SUNDAYTIMES"
 };
 
-storiesOf("DatePublication", module).add("DatePublication", () => (
-  <DatePublication {...props} />
+const styles = StyleSheet.create({
+  text: {
+    color: "#696969",
+    fontSize: 13,
+    lineHeight: 15,
+    fontFamily: "GillSansMTStd-Medium"
+  }
+});
+
+storiesOf("DatePublication", module).add("standard DatePublication", () => (
+  <Text style={styles.text}>
+    <DatePublication {...props} />
+  </Text>
 ));