+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.0"></a>
+# [0.6.0](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.5.2...@times-components/article-image@0.6.0) (2017-12-07)
+
+
+### Features
+
+* make article responsive ([#421](https://github.com/newsuk/times-components/issues/421)) ([3c2aaf7](https://github.com/newsuk/times-components/commit/3c2aaf7))
+
+
+
+
 <a name="0.5.2"></a>
 ## [0.5.2](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.5.0...@times-components/article-image@0.5.2) (2017-12-04)